# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "欺骗防御接口文档",
        "contact": {
            "name": "akita.tian",
            "email": "chronoyq@163.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8082",
    "paths": {
...
        "/api/v1/proxy/protocol": {
            "post": {
                "description": "创建协议代理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "影子代理"
                ],
                "summary": "创建协议代理",
                "parameters": [
                    {
                        "description": "ProxyPort",
                        "name": "ProxyPort",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol_proxy_handler.ProtocolProxyCreatePayload"
                        }
                    },
                    {
                        "description": "HoneypotID",
                        "name": "HoneypotID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol_proxy_handler.ProtocolProxyCreatePayload"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "1007": {
                        "description": "{\"code\":1007,\"msg\":\"Redis异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "6001": {
                        "description": "{\"code\":6001,\"msg\":\"协议代理创建失败\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "6006": {
                        "description": "{\"code\":6006,\"msg\":\"代理端口重复\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/proxy/protocol/:id": {
            "delete": {
                "description": "删除协议代理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "影子代理"
                ],
                "summary": "删除协议代理",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/proxy/protocol/offline/:id": {
            "post": {
                "description": "下线协议代理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "影子代理"
                ],
                "summary": "下线协议代理",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/proxy/protocol/online/:id": {
            "post": {
                "description": "上线协议代理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "影子代理"
                ],
                "summary": "上线协议代理",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/proxy/protocol/set": {
            "post": {
                "description": "查找协议代理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "影子代理"
                ],
                "summary": "查找协议代理",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.SelectPayload"
                        }
                    },
                    {
                        "description": "PageNumber",
                        "name": "PageNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.SelectPayload"
                        }
                    },
                    {
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.SelectPayload"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.ProtocolProxySelectResultPayload"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/proxy/protocol/test/:id": {
            "get": {
                "description": "测试协议代理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "影子代理"
                ],
                "summary": "测试协议代理",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "comm.ProtocolProxySelectResultPayload": {
            "type": "object",
            "properties": {
                "CreateTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "Creator": {
                    "description": "创建用户",
                    "type": "string"
                },
                "HoneypotIP": {
                    "description": "蜜罐IP",
                    "type": "string"
                },
                "HoneypotName": {
                    "description": "蜜罐名称",
                    "type": "string"
                },
                "ID": {
                    "description": "协议代理ID",
                    "type": "integer"
                },
                "MaxPort": {
                    "description": "协议MaxPort",
                    "type": "integer"
                },
                "MinPort": {
                    "description": "协议MinPort",
                    "type": "integer"
                },
                "ProxyPort": {
                    "description": "代理端口",
                    "type": "integer"
                },
                "ServerIP": {
                    "description": "蜜网IP",
                    "type": "string"
                },
                "ServerPort": {
                    "description": "蜜罐端口",
                    "type": "string"
                },
                "ServerType": {
                    "description": "蜜罐服务",
                    "type": "string"
                },
                "Status": {
                    "description": "状态",
                    "type": "integer"
                }
            }
        },
...
        "comm.SelectPayload": {
            "type": "object",
            "required": [
                "PageNumber",
                "PageSize"
            ],
            "properties": {
                "PageNumber": {
                    "description": "页number",
                    "type": "integer"
                },
                "PageSize": {
                    "description": "页size",
                    "type": "integer"
                },
                "Payload": {
                    "description": "查找payload",
                    "type": "string"
                }
            }
        },
...
        "protocol_proxy_handler.ProtocolProxyCreatePayload": {
            "type": "object",
            "required": [
                "HoneypotID",
                "ProxyPort"
            ],
            "properties": {
                "HoneypotID": {
                    "type": "integer"
                },
                "ProxyPort": {
                    "type": "integer"
                }
            }
        },
...
    }
}